+++ src/main/java/me/rrf/web/MVCAdapter.java
@@ -1,4 +1,4 @@
-package me.rrf.caj;
+package me.rrf.web;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
